contributors/ttwf/jackyy/incoming/css-transforms/transform_translate_max_prefixed.html

<!DOCTYPE html>
<html>
<head>
    <title>CSS Transforms API Test: transform translate with maximum translation value</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="Check if transform supports translate(INFINITE, INFINITE)"> 
    <script type="text/javascript" src="../../resources/testharness.js"></script>
    <script type="text/javascript" src="../../resources/testharnessreport.js"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
        // Set the transform 
        document.getElementById("test").style.webkitTransform = "translate(INFINITE, INFINITE)";
        /* document.getElementById("test").style.mozkitTransform = "translate(INFINITE, INFINITE)";
        document.getElementById("test").style.msTransform = "translate(INFINITE, INFINITE)";
        document.getElementById("test").style.oTransform = "translate(INFINITE, INFINITE)";
        document.getElementById("test").style.transform = "translate(INFINITE, INFINITE)"; */
 
        // Verify that the transform was set as expected
        var value = document.getElementById("test").style.getPropertyCSSValue("-webkit-transform").cssText;
        /* var value = document.getElementById("test").style.getPropertyCSSValue("-moz-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("-ms-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("-0-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("transform").cssText; */
        
        test(function() {assert_equals(value,
                        "translate(INFINITE, INFINITE)",
                        "transform should be translate(INFINITE, INFINITE)")}, 
                        "transform_translate_INFINITE_INFINITE");
    </script>    
</body>
</html>
